requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.1.0
backports.tarfile==1.2.0
beautifulsoup4==4.13.3
blinker==1.9.0
boto3==1.37.2
botocore==1.37.2
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
crispy-bootstrap5==2024.10
cryptography==44.0.2
cycler==0.12.1
deepface==0.0.93
distro==1.9.0
Django==5.1.6
django-admin-interface==0.29.4
django-bootstrap5==24.3
django-colorfield==0.12.0
django-crispy-forms==2.3
django-cron==0.6.0
django-crontab==0.7.1
django-extensions==4.1
django-localflavor==4.0
djangorestframework==3.16.1
dlib==20.0.0
docutils==0.21.2
dotenv==0.9.9
exceptiongroup==1.2.2
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.115.8
filelock==3.17.0
fire==0.7.0
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
gast==0.6.0
gdown==5.2.0
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-vision==3.10.0
google-pasta==0.2.0
googleapis-common-protos==1.69.1
groq==0.18.0
grpcio==1.71.0
grpcio-status==1.62.3
gunicorn==23.0.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
id==1.5.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jax==0.5.1
jaxlib==0.5.1
jeepney==0.9.0
Jinja2==3.1.5
jiter==0.10.0
jmespath==1.0.1
joblib==1.4.2
keras==3.8.0
keyring==25.6.0
kiwisolver==1.4.8
libclang==18.1.1
lz4==4.4.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mediapipe==0.10.21
ml-dtypes==0.4.1
more-itertools==10.6.0
mtcnn==1.0.0
multidict==6.1.0
mysqlclient==2.2.7
namex==0.0.8
nh3==0.2.21
numpy==1.26.4
openai==0.28.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pillow==11.1.0
propcache==0.3.0
proto-plus==1.26.0
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyMuPDF==1.26.3
pyparsing==3.2.1
pyproject_hooks==1.2.0
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.20
python-slugify==8.0.4
python-stdnum==1.20
pytz==2025.1
readme_renderer==44.0
reportlab==4.4.0
requests==2.32.3
requests-toolbelt==1.0.0
retina-face==0.0.17
rfc3986==2.0.0
rich==13.9.4
rsa==4.9
s3transfer==0.11.3
schedule==1.2.2
scipy==1.15.2
SecretStorage==3.3.3
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
sqlparse==0.5.3
starlette==0.45.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-intel==0.0.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
text-unidecode==1.3
tomli==2.2.1
tqdm==4.67.1
twine==6.1.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
